move some shit around
